+++ client/vite.config.ts
@@ -2,6 +2,9 @@ import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 import tailwindcss from "@tailwindcss/vite";
 
+// @ts-expect-error process is a nodejs global
+const host = process.env.TAURI_DEV_HOST;
+
 export default defineConfig({
   plugins: [react(), tailwindcss()],
   clearScreen: false,
@@ -9,6 +12,14 @@ export default defineConfig({
   server: {
     port: 1420,
     strictPort: true,
+    host: host || false,
+    hmr: host
+      ? {
+          protocol: "ws",
+          host,
+          port: 1421,
+        }
+      : undefined,
   },
   build: {
     outDir: "dist",